+++ src/main/connections.c
@@ -2273,7 +2273,7 @@ static void raw_resize(Rrawconn this, size_t needed)
     size_t nalloc = 64;
     SEXP tmp;
 
-    if (needed > 8192) nalloc = (size_t)(1.2*needed); /* 20% over-allocation */
+    if (needed > 8192) nalloc = (size_t)(1.2*(double)needed); /* 20% over-allocation */
     else while(nalloc < needed) nalloc *= 2;  /* use powers of 2 if small */
     PROTECT(tmp = allocVector(RAWSXP, nalloc));
     memcpy(RAW(tmp), RAW(this->data), this->nbytes);
